Feature/seller byer change

+++ src/main/java/no/maoyi/app/conversation/entity/ConversationDTO.java
@@ -29,7 +29,7 @@ public class ConversationDTO {
                                               .collect(Collectors.toList());
     }
 
-    BigInteger id;
+    long id;
 
     Listing baseOrder;
